refactor (search) Move apis url handling to api layer
